+++ dropwizard-client/src/main/java/io/dropwizard/client/JerseyClientBuilder.java
@@ -1,6 +1,7 @@
 package io.dropwizard.client;
 
 import com.codahale.metrics.MetricRegistry;
+import com.codahale.metrics.httpclient.HttpClientMetricNameStrategy;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
@@ -14,6 +15,8 @@ import org.apache.http.conn.DnsResolver;
 import org.apache.http.conn.socket.ConnectionSocketFactory;
 import org.glassfish.jersey.apache.connector.ApacheConnectorProvider;
 import org.glassfish.jersey.client.ClientConfig;
+import org.glassfish.jersey.client.ClientProperties;
+import org.glassfish.jersey.client.RequestEntityProcessing;
 
 import javax.validation.Validation;
 import javax.validation.Validator;
@@ -180,6 +183,17 @@ public class JerseyClientBuilder {
     }
 
     /**
+     * Use the given {@link HttpClientMetricNameStrategy} instance.
+     *
+     * @param metricNameStrategy    a {@link HttpClientMetricNameStrategy} instance
+     * @return {@code this}
+     */
+    public JerseyClientBuilder using(HttpClientMetricNameStrategy metricNameStrategy) {
+        builder.using(metricNameStrategy);
+        return this;
+    }
+
+    /**
      * Builds the {@link Client} instance.
      *
      * @return a fully-configured {@link Client}
@@ -234,6 +248,9 @@ public class JerseyClientBuilder {
         for (Map.Entry<String, Object> property : this.properties.entrySet()) {
             config.property(property.getKey(), property.getValue());
         }
+
+        final RequestEntityProcessing requestEntityProcessing = configuration.isChunkedEncodingEnabled() ? RequestEntityProcessing.CHUNKED : RequestEntityProcessing.BUFFERED;
+        config.property(ClientProperties.REQUEST_ENTITY_PROCESSING, requestEntityProcessing);
 
         config.register(new DropwizardExecutorProvider(threadPool));
         config.connectorProvider(new ApacheConnectorProvider());