+++ dropwizard-hibernate/src/test/java/io/dropwizard/hibernate/SessionFactoryManagerTest.java
@@ -25,4 +25,11 @@ public class SessionFactoryManagerTest {
 
         verify(dataSource).stop();
     }
+
+    @Test
+    public void startsTheDataSourceOnStarting() throws Exception {
+        manager.start();
+
+        verify(dataSource).start();
+    }
 }