+++ dropwizard-migrations/src/main/java/io/dropwizard/migrations/CloseableLiquibase.java
@@ -13,14 +13,14 @@ public class CloseableLiquibase extends Liquibase implements AutoCloseable {
     private static final String DEFAULT_MIGRATIONS_FILE = "migrations.xml";
     private final ManagedDataSource dataSource;
 
-    public CloseableLiquibase(ManagedDataSource dataSource) throws LiquibaseException, ClassNotFoundException, SQLException {
+    public CloseableLiquibase(ManagedDataSource dataSource) throws LiquibaseException, SQLException {
         super(DEFAULT_MIGRATIONS_FILE,
               new ClassLoaderResourceAccessor(),
               new JdbcConnection(dataSource.getConnection()));
         this.dataSource = dataSource;
     }
 
-    public CloseableLiquibase(ManagedDataSource dataSource, String file) throws LiquibaseException, ClassNotFoundException, SQLException {
+    public CloseableLiquibase(ManagedDataSource dataSource, String file) throws LiquibaseException, SQLException {
         super(file,
               new FileSystemResourceAccessor(),
               new JdbcConnection(dataSource.getConnection()));