+++ docs/source/about/release-notes.rst
@@ -4,6 +4,30 @@
 Release Notes
 #############
 
+.. _rel-0.8.0:
+
+v0.8.0:
+=======
+
+* Upgrade to argparse4j 0.4.4
+* Upgrade to commons-lang3 3.3.2
+* Upgrade to Guava 18.0
+* Upgrade to H2 1.4.182
+* Upgrade to Hibernate 4.3.7.Final
+* Upgrade to Hibernate Validator 5.1.3.Final
+* Upgrade to Jackson 2.4.3
+* Upgrade to Jersey 2.13
+* Upgrade to Jetty 9.2.4.v20141103
+* Upgrade to Joda-Time 2.5
+* Upgrade to Liquibase 3.3.0
+* Upgrade to Mustache 0.8.16
+* Upgrade to SLF4J 1.7.7
+* Upgrade to tomcat-jdbc 8.0.15
+* Upgrade to JSR305 annotations 3.0.0
+* Upgrade to AssertJ 1.7.0
+* Upgrade to Mockito 1.10.8
+
+
 .. _rel-0.7.1:
 
 v0.7.1: Jun 18 2014
@@ -91,7 +115,7 @@ v0.7.0: Apr 04 2014
   are now excluded by default.
 * Added the application name and PID (if detectable) to the beginning of syslog messages, as is the
   convention.
-* Added ``--migrations-file`` command-line option to ``migrate`` command to supply the migrations
+* Added ``--migrations`` command-line option to ``migrate`` command to supply the migrations
   file explicitly.
 * Validation errors are now returned as ``application/json`` responses.
 * Simplified ``AsyncRequestLog``; now standardized on Jetty 9 NCSA format.